Парсер перенесен в пакет utils
