+++ src/main/java/util/parser/parsetypes/ParseMaterial.java
@@ -1,4 +1,4 @@
-package parser.parsetypes;
+package util.parser.parsetypes;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
